# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "gocommerce",
    "version": "0.1.0",
    "description": "A small, composable commerce engine for Go.\n\nMoney is always an integer count of the currency's minor unit plus its code — never a float, never a formatted string. The number of decimal places belongs to the currency (USD has 2, JPY has 0, KWD has 3) and formatting belongs to the client.\n\nEvery response is `{\"data\": ...}` or `{\"error\": {\"code\", \"message\"}}`. Every collection paginates.\n\nGuest checkout is permanent: a shopper needs a cart token and an email, never an account.",
    "license": {
      "name": "TBD"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "This store"
    }
  ],
...
  "security": [],
  "paths": {
...
    "/api/admin/tax-rates": {
      "get": {
        "tags": [
          "admin:catalog"
        ],
        "summary": "List tax rates, most specific first",
        "description": "The order they come back in is the order they are applied in, so a reader can see which rule wins without reproducing the resolution.",
        "security": [
          {
            "adminToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Rates.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TaxRate"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ListMeta"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "admin:catalog"
        ],
        "summary": "Create a tax rate",
        "description": "One rule per country, state and category. A second rule covering the same three is refused: that is a mistake to catch at entry, not a preference to resolve at checkout.",
        "security": [
          {
            "adminToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaxRateInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created rate.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/TaxRate"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Validation"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/api/admin/tax-rates/{id}": {
      "get": {
        "tags": [
          "admin:catalog"
        ],
        "summary": "Get a tax rate",
        "security": [
          {
            "adminToken": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PathID"
          }
        ],
        "responses": {
          "200": {
            "description": "The rate.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/TaxRate"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "patch": {
        "tags": [
          "admin:catalog"
        ],
        "summary": "Update a tax rate",
        "description": "Changing a rate affects what is charged from now on and nothing already sold. What each order was charged is stored on its lines, so an invoice prints the same figures next year as it does today.",
        "security": [
          {
            "adminToken": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PathID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaxRateInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated rate.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/TaxRate"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Validation"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      },
      "delete": {
        "tags": [
          "admin:catalog"
        ],
        "summary": "Delete a tax rate",
        "description": "Orders keep what they were charged.",
        "security": [
          {
            "adminToken": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/PathID"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted."
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "adminToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Either a static token from Config.AdminTokens — the right credential for a script, since it has no session and no expiry — or a superuser session token from /api/admin/auth-with-password, which is the right credential for a person. Both arrive the same way and satisfy the same middleware."
      }
    },
    "parameters": {
      "PathID": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int64",
          "minimum": 1
        }
      },
...
    },
    "responses": {
      "NotFound": {
        "description": "No such resource.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Validation": {
        "description": "The request was malformed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Conflict": {
        "description": "The request conflicts with the current state.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "A valid admin token is required.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "ErrorEnvelope": {
        "type": "object",
        "description": "Every failure uses this shape, whatever endpoint produced it.",
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "not_found",
                  "method_not_allowed",
                  "validation_failed",
                  "conflict",
                  "unauthorized",
                  "not_ready",
                  "internal_error"
                ]
              },
              "message": {
                "type": "string"
              },
              "details": {
                "description": "Structured detail, such as per-line checkout conflicts."
              }
            }
          }
        }
      },
      "ListMeta": {
        "type": "object",
        "description": "Accompanies every paginated collection. Both coordinate systems are reported: `offset` for a client walking a cursor, and `page` for a UI drawing \"3 of 12\". The page fields are derived from the offset, so the two can never disagree.",
        "required": [
          "total",
          "limit",
          "offset",
          "page",
          "total_pages"
        ],
        "properties": {
          "total": {
            "type": "integer",
            "description": "Matching records in the whole collection, not just this page."
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          },
          "page": {
            "type": "integer",
            "description": "The page this window falls in, counting from 1."
          },
          "total_pages": {
            "type": "integer",
            "description": "Pages at this limit. Zero when nothing matched."
          }
        }
      },
...
      "TaxRate": {
        "type": "object",
        "description": "A rule about a place and a kind of thing. The most specific rule that fits a line is the one that applies: a category beats a country, a country beats a bare fallback, and within categories the deeper one wins. A rule on a category reaches its whole subtree.",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "description": "What appears on the invoice, e.g. \"GST 18%\"."
          },
          "rate_bp": {
            "type": "integer",
            "description": "Basis points: 1800 is 18.00%. An integer, like every other money-adjacent figure here."
          },
          "country": {
            "type": "string",
            "description": "ISO 3166-1 alpha-2. Empty is the fallback every other rule is more specific than."
          },
          "state": {
            "type": "string",
            "description": "Requires a country."
          },
          "category_id": {
            "type": "integer",
            "format": "int64",
            "nullable": true,
            "description": "Absent applies to everything."
          },
          "category_name": {
            "type": "string",
            "description": "Read-only, for display."
          },
          "active": {
            "type": "boolean"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TaxRateInput": {
        "type": "object",
        "required": [
          "name",
          "rate_bp"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "What appears on the invoice, e.g. \"GST 18%\"."
          },
          "rate_bp": {
            "type": "integer",
            "description": "Basis points: 1800 is 18.00%. An integer, like every other money-adjacent figure here."
          },
          "country": {
            "type": "string",
            "description": "ISO 3166-1 alpha-2. Empty is the fallback every other rule is more specific than."
          },
          "state": {
            "type": "string",
            "description": "Requires a country."
          },
          "category_id": {
            "type": "integer",
            "format": "int64",
            "nullable": true,
            "description": "Absent applies to everything."
          },
          "active": {
            "type": "boolean"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
...
    }
  }
}